requirements.txt
discord.py
python-dotenv
aiohttp
beautifulsoup4
playwright
